+++ Vertex.cpp
@@ -3,8 +3,9 @@
 //
 
 #include "Vertex.h"
+#include "City.h"
 
-Vertex::Vertex(int x, int y, int id) {
+Vertex::Vertex(float x, float y, int id) {
     set_x(x);
     set_y(y);
     set_id(id);
